requirements_test.txt
aresponses==2.1.4
coverage==5.5
flake8==3.9.2
flake8-docstrings==1.6.0
isort==5.9.1
mypy==0.790
pylint==2.9.3
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-cov==2.12.1
pytest-aiohttp=0.3.0
